add ece and move earlystopping to utils
